split the flask application using blueprints for better code management
